+++ clang/docs/ReleaseNotes.rst
@@ -397,6 +397,7 @@ clang-format
 
 libclang
 --------
+- visit identifier initializers in lambda capture as VarDecl
 
 Code Completion
 ---------------